Use logit_dtype for validation loss

+++ scripts/train.py
@@ -33,7 +33,8 @@ def train(model,
         if epoch % validate_every == 0:
             model.eval()
             with torch.inference_mode():
-                loss_valid = sum(loss_fn(model(xb), yb) for xb, yb in dl_valid) / len(dl_valid)
+                loss_valid = sum(loss_fn(model(xb).to(logit_dtype), yb)
+                                 for xb, yb in dl_valid) / len(dl_valid)
             wandb.log({"loss_valid": loss_valid.item()}, commit=False)
             checkpoint_name = f"checkpoint-{epoch}.pt"
             checkpoint_path = os.path.join(wandb.run.dir, checkpoint_name)
